[IMP] *: don't optionally import non-optional requirements
- openssl is a hard dependency of base (ir_mail_server)
- cryptography is a dependency of openssl
- stdnum is a hard dependency of `partner_autocomplete`, which is auto_install

Also move the patching of xlsxwriter and xlwt to the lazy
monkeypatcher, and stop importing xlsxwriter from the patch
location (where it was unwittingly exposed as a side effect).
