+++ addons/web/controllers/pivot.py
@@ -1,16 +1,18 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from collections import deque
 import io
 import json
+from collections import deque
 
 from werkzeug.datastructures import FileStorage
+try:
+    import xlsxwriter
+except ImportError:
+    xlsxwriter = None
 
 from odoo import http, _
 from odoo.http import content_disposition, request
 from odoo.tools import osutil
-from odoo.tools.misc import xlsxwriter
 
 
 class TableExporter(http.Controller):
