+++ addons/web/models/base_document_layout.py
@@ -3,7 +3,6 @@ import os
 from markupsafe import Markup
 
 from odoo import api, fields, models
-
 from odoo.addons.base.models.ir_qweb_fields import nl2br
 from odoo.tools import html2plaintext, is_html_empty, image as tools
 from odoo.tools.misc import file_path
