+++ addons/web/models/res_partner.py
@@ -1,6 +1,5 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
-
 import logging
 from base64 import b64decode
 
@@ -9,7 +8,7 @@ from odoo import models
 _logger = logging.getLogger(__name__)
 
 try:
-    import vobject
+    import vobject.vcard
 except ImportError:
     _logger.warning("`vobject` Python module not found, vcard file generation disabled. Consider installing this module if you want to generate vcard files")
     vobject = None
