+++ odoo/_monkeypatches/__init__.py
@@ -75,6 +75,6 @@ def patch_init():
     patch_module('win32')
     for name in (
         'ast', 'email', 'mimetypes', 'num2words', 'pytz', 'stdnum', 'urllib3',
-        'werkzeug', 'zeep',
+        'werkzeug', 'xlsxwriter', 'xlwt', 'zeep',
     ):
         HOOK_IMPORT.add_hook(name, functools.partial(patch_module, name))
